feat: added the details modal
